roshni_dholariya/requirements.txt
streamlit
scikit-learn
pandas
numpy
sentence-transformers
python-docx
pdfminer.six
joblib
